requirements.txt
selenium
requests
pillow
tqdm